\name{Bst.permutation.plot}
\alias{Bst.permutation.plot}
\title{plotting observed and expected Bst across branching times of a phylogeny}
\usage{
  Bst.permutation.plot(Bst.permutations, cex = list(1.6, 0.1, 0.6), col = list("black", "lightgray"), 
  bg = list("white", "lightgray", "black"), all.points = TRUE, add.id = TRUE, sig.plot = TRUE, cut.off = 0.05, envelope = FALSE, outfile = NULL, ...)
}

\arguments{
  \item{Bst.permutations}{a list-object with at least two elements observed and expected \code{Bst}s, 
  typically generated by \code{\link{Bst.by.nodes}}} 
  \item{cex}{a list of three elements: size of points for observed \code{Bst}s; for expected \code{Bst}s; and for size of text for \code{node labels}}
  \item{col}{a list of two elements: colors to be used for plotting observed \code{Bst}s and expected \code{Bst}s}
  \item{bg}{a list of two or three elements: background (fill) colors to be used for plotting observed \code{Bst}s and expected \code{Bst}s}
  \item{all.points}{a logical indicator determining whether all expected \code{Bst}s are to be plotted}
  \item{add.id}{a logical indicator dictating whether \code{node.labels} are placed near observed \code{Bst}s; see \code{\link[ape]{nodelabels}}}
  \item{sig.plot}{a logical indicator dictating whether nodes are colored by significance of observed and expected \code{Bst}s}
  \item{cut.off}{a value, if \code{sig.plot=TRUE}, distinguishing observed \code{Bst}s from expected \code{Bst}s}
  \item{envelope}{a logical indicator specifying whether a smooth-spline envelope is drawn around expected \code{Bst}s}
  \item{outfile}{an optional .pdf file to which to write output}
  \item{\dots}{additional plotting parameters to be optionally supplied}
}

\details{
  This function generates a plot of phylogenetic turnover (\code{Bst}), evaluated for each subtree of the supplied phylogeny. 
  If \code{all.points=TRUE}, empirical \code{Bst}s as well as those expected from permuted community datasets are plotted together. 
  Several methods are available for data permutation when using \code{\link{Bst.by.nodes}}, typically the function used 
  to generate \code{Bst.permutations}. If \code{envelope=TRUE} an envelope for a credible range of \code{Bst}s is additionally plotted.  
}

\value{a plot of \code{Bst} values for a phylogeny, with the option to direct output to a .pdf file}

\author{Jonathan Eastman}


\seealso{\code{\link{spacodi.calc}} for further details on \code{Bst}; \code{\link{Bst.by.nodes}}; \code{\link[stats]{smooth.spline}}}

\examples{
# load a species-by-plots matrix, along with a tree
data(sp.example)
attach(sp.example)
sp.plot
sp.tree

# generate a plot of observed and expected Bst
Bst.permutation.plot(Bst.by.nodes(phy=sp.tree, sp.plot=sp.plot, n.rep=10)->sp.permut, cex=list(2,0.25,0.75), bg=list("white",1), col=list(1,1),add.id=TRUE, envelope=FALSE, lwd=0.4)
}