\name{Bst.permutation}

\alias{Bst.permutation}

\title{conducting permutation tests of phylogenetic turnover}

\usage{Bst.permutation(phy, sp_plot, n.rep = 10, method = "1s", parm = NULL)}

\arguments{
  \item{phy}{a fully resolved phylogenetic tree, most sensibly ultrametric}
  \item{sp_plot}{a species-by-plots matrix or dataframe}
  \item{n.rep}{number of iterations to perform}
  \item{method}{reshuffling or simulation procedure to perform (see details for optional methods)}
  \item{parm}{additional parameters, required for particular methods: \code{1a, 2x, and 3x} (see details)}
}

\details{
  Bst is a measure of local phyletic proximity excess between individuals of distinct species, 
  thereby expressing phylogenetic turnover.  This measure considers the balance between 
  phylogenetic diversity within and among plots. 
    
  \code{Bst.permutation} computes phylogenetic community diversity for nodes of 
  a phylogeny that subtend greater than two species and where nodes subtend species 
  that are sampled in multiple plots. 
    
  For significance testing, permuted datasets may be generated by one of several \code{methods}:
    \itemize{
	\item \code{method = "1a"} shuffling based on abundance classes of species (see \code{\link{resamp.1a}}); 
	requires specification of \code{abund.class.ratio} as a \code{parm} (see examples)
	
	\item \code{method = "1s"} shuffling of abundances across entire dataset (see \code{\link{resamp.1s}})

	\item \code{method = "2s"} shuffling of abundances across species but within plots (see \code{\link{resamp.2s}})
	
	\item \code{method = "2x"} Gotelli swapping of abundances for pairs of species and within a pair of plots (see \code{\link{resamp.2x}}); 
	requires specification of \code{level} as a \code{parm} (see examples)
	
	\item \code{method = "3i"} shuffling of abundances within species and among plots (see \code{\link{resamp.3i}})
	
	\item \code{method = "3t"} shuffling of abundances to adjacent plots but within species (see \code{\link{resamp.3t}})

	\item \code{method = "3x"} Gotelli swapping of abundances for pairs of plots and within a pair of species (see \code{\link{resamp.3x}}); 
	requires specification of \code{level} as a \code{parm} (see examples)
	
	\item \code{method = "r.sp_plot"} simulation of species-by-plots matrices, informed by empirical data (see \code{\link{r.sp_plot}})
  }
}

\value{  
  \code{Bst.permutation} returns an array of community diversity measures (Bst) 
  from permuted datasets.  The first dimension of the array is determined
  by is the number of internal nodes of the supplied phylogeny. The second dimension 
  of the output array are the measures returned to the user: \code{Bst} -- computed for each node (if nodes 
  subtend more than two sampled species and if subtended species were sampled in multiple 
  plots); \code{tips} -- the number of contained species at each node; \code{node.time} -- the branching time of the
  node in the supplied phylogeny (in same units as those supplied); and \code{node.ID} -- the identifier for 
  each node (as interpreted by \code{ape:::nodelabels()}). The third dimension of the output array is specified by \code{n.rep}.
  
  To access output from \code{Bst.permutation}, the stored object (e.g., as a \code{out = Bst.permutation(phy, sp_plot)})
  is indexed by the three dimensions of the array (as above).  For instance, output from the third iteration 
  would be found using \code{out[ , ,3]}.  Analogously, all \code{Bst} and \code{node.time} output (i.e., across all iterations) 
  for the entire phylogeny (i.e., root node only) would be found by \code{out[1, c(1,3), ]}. 
}

\references{
  HARDY OJ and B SENTERRE. 2007. Characterizing the 
  phylogenetic structure of communities by an additive partitioning of 
  phylogenetic diversity. Journal of Ecology 95:493-506.
  
  HARDY OJ. 2008. Testing the spatial phylogenetic 
  structure of local communities: statistical performances of 
  different null models and test statistics on a locally neutral 
  community. Journal of Ecology 96:914-926.
}

\author{Timothy Paine and Jonathan Eastman}

\seealso{\code{\link{spacodi.calc}} for interpretation of results; see \code{\link{Bst.all.nodes}} for computing \code{Bst} for the empirical dataset.}

\examples{
# generate a random dataset and phylogeny 
foo <- r.plot(species=10,plots=4,missing.prop=0.15,sim.tree=TRUE)
sp_plot <- foo[[1]]
phy <- foo[[2]]

# shuffle dataset by '1a' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="1a", parm=list(abund.class.ratio=2))
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

# shuffle dataset by '1s' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="1s")
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

# shuffle dataset by '2s' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="2s")
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

# shuffle dataset by '2x' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="2x", parm=list(level=0.05))
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

# shuffle dataset by '3i' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="3i")
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

# shuffle dataset by '3t' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="3t")
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

# shuffle dataset by '3x' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="3x", parm=list(level=0.05))
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

# shuffle dataset by 'r.sp_plot' permutations
Bst.permutation(sp_plot=sp_plot, phy=phy, n.rep=2, method="r.sp_plot")
Bst.all.nodes(phy, sp_plot) ## compare above with EMPIRICAL dataset

}
