\name{K.simulation.plot}
\alias{K.simulation.plot}
\title{plotting observed and expected K across branching times of a phylogeny}
\usage{
  K.simulation.plot(K.simulations, cex = list(1.6, 0.1, 0.6), col = list("black", "lightgray"), 
  bg=list("white", "lightgray", "black"), all.points = TRUE, add.id = TRUE, sig.plot = TRUE, cut.off = 0.05, envelope = FALSE, outfile = NULL, ...)
}

\arguments{
  \item{K.simulations}{a list-object with at least two elements observed and expected \code{K}s, 
  typically generated by \code{\link{K.by.nodes}}} 
  \item{cex}{a list of three elements: size of points to be used to plot observed \code{K}s; for expected \code{K}s; and for size of text for \code{node labels}}
  \item{col}{a list of two elements: (border) colors to be used for plotting observed \code{K}s and expected \code{K}s}
  \item{bg}{a list of three elements: background (fill) colors to be used for plotting observed \code{K}s and expected \code{K}s. The first color is for nodes for which 
  \code{K} does not depart from expectations; the second and third colors are for \code{K}s that are smaller and larger than expected, respectively}
  \item{all.points}{a logical indicator determining whether all expected \code{K}s are to be plotted. Only the first two elements of the list are used if \code{sig.plot=FALSE}}
  \item{add.id}{a logical indicator dictating whether \code{node.labels} are placed near observed \code{K}s; see \code{\link[ape]{nodelabels}} and \code{\link[calibrate]{textxy}}}
  \item{sig.plot}{a logical indicator dictating whether nodes are colored by significance of observed and expected \code{K}s}
  \item{cut.off}{a value, if \code{sig.plot=TRUE}, distinguishing observed \code{K}s from expected \code{K}s}
  \item{envelope}{a logical indicator specifying whether a smooth-spline envelope is drawn around expected \code{K}s}
  \item{outfile}{an optional .pdf file to which to write output}
  \item{\dots}{additional plotting parameters to be optionally supplied}
}

\details{
  This function generates a plot of phylogenetic trait heritability (\code{K}), evaluated for each subtree (i.e., each node) of the supplied phylogeny. 
  If \code{all.points=TRUE}, empirical \code{K}s as well as those expected from simulated datasets are plotted together. 
}

\value{a plot of \code{K} values for a phylogeny, with the option to direct output to a .pdf file}

\author{Jonathan Eastman}


\seealso{\code{\link[picante]{Kcalc}} for further details on \code{K}; \code{\link{K.by.nodes}}; \code{\link[stats]{smooth.spline}}}

\examples{
# load a species-by-plots matrix, along with a tree
data(sp.example)
attributes(sp.example)
attach(sp.example)
sp.traits
sp.tree

# define a single trait
trt=as.data.frame(sp.traits[,1])
row.names(trt)=row.names(sp.traits)
trt

# generate a plot of observed and expected K
K.simulation.plot(K.by.nodes(sp.traits=trt, phy=sp.tree, n.rep=10)->sp.permut, cex=list(2,0.25,0.75), bg=list("white",1), col=list(1,1),add.id=TRUE, envelope=FALSE, lwd=0.4)
}