\name{sp.example}
\alias{sp.example}
\title{example data for SPACoDi}
\usage{data(sp.example)}
\usage{attach(sp.example)}
\details{This object is a list of example datasets for \pkg{spacodi}: a community-sampling dataset (\code{sp.plot}), 
phylogeny (\code{sp.tree}), species-level traits (\code{sp.traits}), and matrix of distances between plots (\code{sp.dist}. 
The data can be called into R by \code{data(sp.example}, where each item is indexed as an element of a list (e.g., for the 
tree, use \code{sp.example$sp.tree}.}
