\name{spacodi-internal}
\alias{nodelabels.sp}
\alias{BOTHlabels.sp}

\title{internal SPACoDi functions}

\details{
This is an internal \pkg{spacodi} function, not intended to be called directly by the user. \code{nodelabels.sp} is a slight modification of 
\code{\link[ape]{nodelabels}}. \code{BOTHlabels.sp} is a slight modification of \code{\link[ape]{BOTHlabels}}.
}